Add styling to primevue multiselect and deprecate forge multiselect
