+++ apps/docs/src/pages/forms/dropdowns/Multiselect.vue
@@ -1,84 +1,74 @@
 <template>
   <div>
     <ForgePageHeader title="Multiselect" />
-    <p>
-      Further documentation and examples can be found in the
-      <a class="link" target="_blank" href="https://github.com/shentao/vue-multiselect/tree/next?tab=readme-ov-file#vue-multiselect">
-        <strong>Vue Multiselect documentation</strong>
+    <p>Further documentation and examples can be found in the
+      <a class="link" target="_blank" href="https://v3.primevue.org/multiselect/">
+        <strong>PrimeVue Multiselect documentation</strong>
       </a>
       .
     </p>
     <Playground :options="options" :code="code" :config="config" @reset="reset">
       <template #component>
-        <component :is="ForgeMultiSelect" v-bind="options" v-model="selected" />
+        <component
+          :is="Multiselect"
+          v-bind="options"
+          v-model="selectedOptions"
+          class="w-100"
+        />
       </template>
     </Playground>
-    Here is an example of a multiselect options array
-    <code-block :code="optionsArrayExample" />
+
+    <h3>Props Info:</h3>
+    <ul>
+      <li>maxSelectedLabels - Limit the selected labels shown, if the number select goes over this the number selected will be shown and not the pills.</li>
+      <li>selectionLimit - Maximum number of selectable items.</li>
+    </ul>
   </div>
 </template>
 
 <script setup lang="ts">
-import { ForgePageHeader, ForgeMultiSelect } from "@3squared/forge-ui-3";
-import { Playground, usePlayground, CodeBlock } from "@3squared/forge-playground-3";
+import { ForgePageHeader,
+  } from "@3squared/forge-ui-3";
+import Multiselect from "primevue/multiselect";
+import { Playground, usePlayground } from "@3squared/forge-playground-3";
 import { computed, ref } from "vue";
-import { severities } from "../../../composables/playgroundOptions";
 
-const multiSelectOptions = [
-  { id: "option-1", label: "Option 1" },
-  { id: "option-2", label: "Option 2" },
-  { id: "option-3", label: "Option 3" },
-  { id: "option-4", label: "Option 4" },
-  { id: "option-5", label: "Option 5" },
-  { id: "option-6", label: "Option 6" }
-];
-const selected = ref([]);
+const selectedOptions = ref([]);
 
-const tagPositions = ["top", "bottom"];
+const optionsArray = ref([
+  { id: 1, label: "Option 1", value: "opt_1" },
+  { id: 2, label: "Option 2", value: "opt_2" },
+  { id: 3, label: "Option 3", value: "opt_3" },
+  { id: 4, label: "Option 4", value: "opt_4" },
+  { id: 5, label: "Option 5", value: "opt_5" },
+  { id: 6, label: "Option 6", value: "opt_6" },
+  { id: 7, label: "Option 7", value: "opt_7" },
+  { id: 8, label: "Option 8", value: "opt_8" },
+  { id: 9, label: "Option 9", value: "opt_9" },
+  { id: 10, label: "Option 10", value: "opt_10" },
+  { id: 11, label: "Option 11", value: "opt_11" },
+  { id: 12, label: "Option 12", value: "opt_12" },
+]);
 
+const placeHolder = ref("Select options");
+
+const displays = ["chip", "comma"];
+
+const code = computed(
+  () =>
+    `<Multiselect ${propVals.value.length > 0 ? " " + propVals.value.join(" ") : ""} />`,
+);
 const { options, propVals, config, reset } = usePlayground(
   {
-    options: multiSelectOptions,
-    severity: severities[0],
-    expandLeft: false,
-    showClearSelection: false,
-    showSelectAll: false,
-    selectValue: "",
-    name: "",
-    label: "label",
-    searchable: true,
-    clearOnSelect: true,
-    hideSelected: false,
-    placeholder: "Search",
-    allowEmpty: true,
-    resetAfter: false,
-    closeOnSelect: false,
-    taggable: false,
-    tagPlaceholder: "Please enter to create a tag",
-    tagPosition: tagPositions[0],
-    max: 9999,
-    optionsLimit: 1000,
-    groupValues: "",
-    groupLabel: "",
-    groupSelect: false,
-    internalSearch: true,
-    preserveSearch: false,
-    preselectFirst: false
+    options: optionsArray,
+    optionLabel: "label",
+    placeholder: placeHolder,
+    display: "comma",
+    maxSelectedLabels: 4,
+    filter: false
   },
   {
-    severity: { type: "select", options: severities },
-    tagPosition: { type: "select", options: tagPositions }
-  }
-);
-
-const code = computed(() => `<ForgeMultiSelect${propVals.value.length > 0 ? " " + propVals.value.join(" ") : ""} />`);
-
-const optionsArrayExample = computed(
-  () => `<script setup lang='ts'>
- const optionsArray = ref([
-   { id: 1, label: 'option 1', value: 'opt_1'},
-   { id: 2, label: 'option 2', value: 'opt_2'}
- ])
-<\/script>`
+    display: { type: "select", options: displays }
+  },
 );
 </script>
