+++ packages/ui/components.d.ts
@@ -30,7 +30,7 @@ declare module 'vue' {
     ForgeLink: typeof import('./src/components/ForgeLink.vue')['default']
     ForgeLoader: typeof import('./src/components/ForgeLoader.vue')['default']
     ForgeModal: typeof import('./src/components/ForgeModal.vue')['default']
-    ForgeMultiSelect: typeof import('./src/components/ForgeMultiSelect.vue')['default']
+    ForgeMultiSelect: typeof import('./src/components/deprecated/ForgeMultiSelect.vue')['default']
     ForgeMultiSelectPreview: typeof import('./src/components/ForgeMultiSelectPreview.vue')['default']
     ForgeNavbar: typeof import('./src/components/ForgeNavbar.vue')['default']
     ForgePageHeader: typeof import('./src/components/ForgePageHeader.vue')['default']
@@ -47,11 +47,15 @@ declare module 'vue' {
     MaxFileSize: typeof import('./src/components/file-uploader/components/MaxFileSize.vue')['default']
     Menubar: typeof import('primevue/menubar')['default']
     Message: typeof import('primevue/message')['default']
+    MultiSelect: typeof import('primevue/multiselect')['default']
+    OverlayPanel: typeof import('primevue/overlaypanel')['default']
     ProgressBar: typeof import('primevue/progressbar')['default']
     ProgressSpinner: typeof import('primevue/progressspinner')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
     Steps: typeof import('primevue/steps')['default']
+    TabPanel: typeof import('primevue/tabpanel')['default']
+    TabView: typeof import('primevue/tabview')['default']
     Textarea: typeof import('primevue/textarea')['default']
     UploadButton: typeof import('./src/components/file-uploader/components/UploadButton.vue')['default']
     UploadStatus: typeof import('./src/components/file-uploader/components/UploadStatus.vue')['default']
